+++ src/main/java/qna/domain/answer/Answer.java
@@ -1,5 +1,6 @@
 package qna.domain;
 
+import org.hibernate.annotations.Where;
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
 
@@ -8,6 +9,7 @@ import java.util.Objects;
 
 @Table(name = "answer")
 @Entity
+@Where(clause = "deleted = false")
 public class Answer extends CreateAndUpdateTimeEntity {
 
     @Id
