+++ src/main/java/qna/domain/history/ContentType.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.history;
 
 public enum ContentType {
     QUESTION, ANSWER
