+++ src/test/java/qna/domain/AnswerTest.java
@@ -1,5 +1,7 @@
 package qna.domain;
 
+import qna.domain.answer.Answer;
+
 public class AnswerTest {
     public static final Answer A1 = new Answer(UserTest.JAVAJIGI, QuestionTest.Q1, "Answers Contents1");
     public static final Answer A2 = new Answer(UserTest.SANJIGI, QuestionTest.Q1, "Answers Contents2");
