+++ src/test/java/qna/domain/QuestionTest.java
@@ -1,5 +1,7 @@
 package qna.domain;
 
+import qna.domain.question.Question;
+
 public class QuestionTest {
     public static final Question Q1 = new Question("title1", "contents1").writeBy(UserTest.JAVAJIGI);
     public static final Question Q2 = new Question("title2", "contents2").writeBy(UserTest.SANJIGI);
