+++ src/test/java/qna/service/QnaServiceTest.java
@@ -8,6 +8,12 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import qna.CannotDeleteException;
 import qna.domain.*;
+import qna.domain.answer.Answer;
+import qna.domain.answer.AnswerRepository;
+import qna.domain.history.ContentType;
+import qna.domain.history.DeleteHistory;
+import qna.domain.question.Question;
+import qna.domain.question.QuestionRepository;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
@@ -88,7 +94,7 @@ class QnAServiceTest {
         List<DeleteHistory> deleteHistories = Arrays.asList(
                 new DeleteHistory(ContentType.QUESTION, question.getId(), question.getWriter(), LocalDateTime.now()),
                 new DeleteHistory(ContentType.ANSWER, answer.getId(), answer.getWriter(), LocalDateTime.now())
-        );
+                                                           );
         verify(deleteHistoryService).saveAll(deleteHistories);
     }
 }
